# webserver/python-flask-server/swagger_server/models/subscription.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Subscription(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, subscriber: str=None, day: int=None):  # noqa: E501
        """Subscription - a model defined in Swagger

        :param subscriber: The subscriber of this Subscription.  # noqa: E501
        :type subscriber: str
        :param day: The day of this Subscription.  # noqa: E501
        :type day: int
        """
        self.swagger_types = {
            'subscriber': str,
            'day': int
        }

        self.attribute_map = {
            'subscriber': 'subscriber',
            'day': 'day'
        }

        self._subscriber = subscriber
        self._day = day

    @classmethod
    def from_dict(cls, dikt) -> 'Subscription':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Subscription of this Subscription.  # noqa: E501
        :rtype: Subscription
        """
        return util.deserialize_model(dikt, cls)

    @property
    def subscriber(self) -> str:
        """Gets the subscriber of this Subscription.


        :return: The subscriber of this Subscription.
        :rtype: str
        """
        return self._subscriber

    @subscriber.setter
    def subscriber(self, subscriber: str):
        """Sets the subscriber of this Subscription.


        :param subscriber: The subscriber of this Subscription.
        :type subscriber: str
        """

        self._subscriber = subscriber

    @property
    def day(self) -> int:
        """Gets the day of this Subscription.


        :return: The day of this Subscription.
        :rtype: int
        """
        return self._day

    @day.setter
    def day(self, day: int):
        """Sets the day of this Subscription.


        :param day: The day of this Subscription.
        :type day: int
        """

        self._day = day