# webserver/python-flask-server/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is the distribution server for recipes and meal plans."
  version: "1.0.0"
  title: "Recipes"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "recipe-dev@yahoo.de"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/v1"
...
schemes:
- "https"
- "http"
paths:
...
  /plan:
    get:
      tags:
      - "plan"
      summary: "Get list of available meal plans"
      description: "Here we return a list of all available meal plans on the server.\
        \ The list itself only contains ID and name of the plan. Use the ID to get\
        \ all details about the plan."
      operationId: "get_plan_list"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Planlist"
      x-swagger-router-controller: "swagger_server.controllers.plan_controller"
    post:
      tags:
      - "plan"
      summary: "Add a plan"
      description: "Use this to add a meal plan to the list of available meal plans\
        \ on the server"
      operationId: "add_plan"
      consumes:
      - "application/json"
      produces:
      - "appliaction/json"
      parameters:
      - in: "body"
        name: "plan"
        description: "Meal plan to add to the server"
        required: true
        schema:
          $ref: "#/definitions/Plan"
      responses:
        "200":
          description: "Successful operation"
        "405":
          description: "Invalid input"
      security:
      - recipe_auth: []
      x-swagger-router-controller: "swagger_server.controllers.plan_controller"
  /plan/{planId}:
    get:
      tags:
      - "plan"
      summary: "Retrieve a meal plan"
      description: "Retrieve detailed information about a single meal plan."
      operationId: "get_plan"
      produces:
      - "application/json"
      parameters:
      - name: "planId"
        in: "path"
        description: "ID of meal plan to return"
        required: true
        type: "string"
        format: "uuid"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Plan"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Plan not found"
      x-swagger-router-controller: "swagger_server.controllers.plan_controller"
    put:
      tags:
      - "plan"
      summary: "Change an existing meal plan"
      description: "Here we change a meal plan that already exists on the server."
      operationId: "change_plan"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "planId"
        in: "path"
        description: "ID of meal plan to change"
        required: true
        type: "string"
        format: "uuid"
      - in: "body"
        name: "plan"
        description: "The changed meal plan"
        required: true
        schema:
          $ref: "#/definitions/Plan"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Plan"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Plan not found"
      security:
      - recipe_auth: []
      x-swagger-router-controller: "swagger_server.controllers.plan_controller"
    delete:
      tags:
      - "plan"
      summary: "Deletes a meal plan"
      description: "Here a meal plan is deleted from the server"
      operationId: "delete_plan"
      produces:
      - "application/json"
      parameters:
      - name: "planId"
        in: "path"
        description: "Id of meal plan to delete"
        required: true
        type: "string"
        format: "uuid"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Planlist"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Plan not found"
      security:
      - recipe_auth: []
      x-swagger-router-controller: "swagger_server.controllers.plan_controller"
  /plan/subscribe/{planId}:
    put:
      tags:
      - "plan"
      summary: "Subscribe to a day in a meal plan"
      description: "Subscribe"
      operationId: "subscribe"
      produces:
      - "application/json"
      parameters:
      - name: "planId"
        in: "path"
        description: "Id of the meal plan to subscribe to"
        required: true
        type: "string"
        format: "uuid"
      - in: "body"
        name: "subscriber"
        description: "The name of the person subscribing to the meal"
        required: true
        schema:
          $ref: "#/definitions/Subscription"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Plan"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Plan not found"
      x-swagger-router-controller: "swagger_server.controllers.plan_controller"
  /plan/unsubscribe/{planId}:
    put:
      tags:
      - "plan"
      summary: "Unsubscribe from a day in a meal plan"
      description: "Unsubscribe"
      operationId: "unsubscribe"
      produces:
      - "application/json"
      parameters:
      - name: "planId"
        in: "path"
        description: "Id of the meal plan to unsubscribe from"
        required: true
        type: "string"
        format: "uuid"
      - in: "body"
        name: "subscriber"
        description: "The name of the person unsubscribing from the meal"
        required: true
        schema:
          $ref: "#/definitions/Subscription"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Plan"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Plan not found"
      x-swagger-router-controller: "swagger_server.controllers.plan_controller"
...
securityDefinitions:
  recipe_auth:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  Planlist:
    type: "array"
    items:
      type: "object"
      properties:
        title:
          type: "string"
        id:
          type: "string"
          format: "uuid"
  Plan:
    type: "object"
    properties:
      name:
        type: "string"
      id:
        type: "string"
        format: "uuid"
      meals:
        type: "integer"
        format: "int32"
      days:
        type: "integer"
        format: "int32"
      eaters:
        type: "array"
        items:
          type: "string"
      items:
        type: "array"
        items:
          type: "object"
          properties:
            shoppingBefore:
              type: "boolean"
            recipes:
              type: "array"
              items:
                type: "object"
                properties:
                  name:
                    type: "string"
                  id:
                    type: "string"
                    format: "uuid"
            subscribers:
              type: "array"
              items:
                type: "string"
    example:
      name: "name"
      eaters:
      - "eaters"
      - "eaters"
      days: 6
      id: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
      items:
      - "{}"
      - "{}"
      meals: 0
  Subscription:
    type: "object"
    properties:
      subscriber:
        type: "string"
      day:
        type: "integer"
        format: "int32"
...